Added player complexity, worked on bee AI
the bulk of the bee ai is done, with attacking, roaming, chasing all implimented (searching for player is not finalized), player now can make their sprite attack and enter/exit stealth but it doesn't have an effect yet. 
To be implemented: attacking will enable/disable a trigger collider connected to the player, if the dot product of the direction towards the mouse and the direction toward colliding enemies is more than a specified value, then it hits. 
stealth will make the screen darker and means bees will not recognise you unless you get close.