Added attack effect, added shaders, worked towards a data storage system
shaders were easy because raylib does half the work. on hit effect took way longer then it should have. floor data is created, in a way similar to binding of isaac. rooms data will be stored in ram, and will be used to generate scenes the same way every time when scenes are switched, as well as to store the connections between each room